rename arm sim to coral arm sim
